01_10/requirements.txt
PyQt6==6.7.1
mysql-connector-python==8.2.0
